# brokerage-apis/openapi/brokerage-api.yaml
openapi: 3.2.1
...
info:
  title: Brokerage API
  version: "2.0"
  summary: Accounts, positions, market data and trading for a retail brokerage, as hypermedia.
  description: |
    A resource-oriented REST API that follows the standards in *REST API Standards with
    HATEOAS* (Smiroh Publishers).

    * **Hypermedia.** Every representation is HAL (`application/hal+json`) with a `_links`
      object. Ask for `application/prs.hal-forms+json` to also receive `_templates`: the
      state transitions available right now. Clients should bookmark only the API root and
      follow links from there. Custom link relations use the CURIE `bk:`, which expands to
      `https://docs.brokerage.example/rels/{rel}`.
    * **Money.** Prices, amounts and quantities are decimal strings, never JSON numbers.
    * **Errors.** Every error is an RFC 9457 problem document (`application/problem+json`)
      whose `type` URI identifies the error. Branch on `type`, never on `title` or `detail`.
    * **Versions.** Send `API-Version: 2` to opt in to version 2. Version 1 is the default and
      is deprecated: its responses carry `Deprecation` (RFC 9745) and `Sunset` (RFC 8594).
    * **Quotas.** Every response carries `RateLimit-Policy` and `RateLimit`. Exceeding a quota
      returns 429 with `Retry-After`.
    * **Tracing.** Every response echoes `X-Request-Id`; quote it in support requests.
  contact:
    name: Brokerage API support
    url: https://docs.brokerage.example/support
  license:
...
servers:
...
paths:
...
  /instruments:
    get:
      operationId: searchInstruments
      summary: Search instruments
      tags: [market-data]
      security: [{ oauth2: [market-data:read] }]
      parameters:
        - name: symbol
          in: query
          schema: { type: string, examples: [AAPL] }
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/InstrumentType" }
      responses:
        "200":
          description: Matching instruments.
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/InstrumentCollection" }
  /instruments/{instrumentId}:
    parameters:
      - $ref: "#/components/parameters/instrumentId"
    get:
      operationId: getInstrument
      summary: Get an instrument
      tags: [market-data]
      security: [{ oauth2: [market-data:read] }]
      responses:
        "200":
          description: The instrument, with links to its quote and option chain or underlying.
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/Instrument" }
        "404": { $ref: "#/components/responses/NotFound" }
  /instruments/{instrumentId}/quote:
    parameters:
      - $ref: "#/components/parameters/instrumentId"
    get:
      operationId: getQuote
      summary: Get a quote
      description: >-
        The shape depends on the API version: version 1 (the deprecated default) is flat;
        version 2 groups price and size into bid and ask levels and adds the day's change.
      tags: [market-data]
      security: [{ oauth2: [market-data:read] }]
      parameters:
        - $ref: "#/components/parameters/API-Version"
        - $ref: "#/components/parameters/If-None-Match"
      responses:
        "200":
          description: The latest quote. Cacheable privately for one second.
          headers:
            Cache-Control:
              schema: { type: string, examples: ["max-age=1, private"] }
            ETag: { $ref: "#/components/headers/ETag" }
            Deprecation: { $ref: "#/components/headers/Deprecation" }
            Sunset: { $ref: "#/components/headers/Sunset" }
          content:
            application/hal+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Quote"
                  - $ref: "#/components/schemas/QuoteV1"
        "304":
          description: The quote has not changed.
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /instruments/{instrumentId}/option-chain:
    parameters:
      - $ref: "#/components/parameters/instrumentId"
    get:
      operationId: getOptionChain
      summary: Get an option chain
      description: >-
        Contracts for one expiration (the nearest by default). Every available expiration is
        linked with the relation `bk:expiration`, named by its date.
      tags: [market-data]
      security: [{ oauth2: [market-data:read] }]
      parameters:
        - name: expiration
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: The option contracts.
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/InstrumentCollection" }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: >-
        OAuth 2.1. Web apps sign customers in through their backend, a confidential client,
        with the authorization code grant and PKCE; a customer's own scripts and agents use
        client credentials registered to that customer. Tokens carry `customer_id` and, for
        delegated clients, `trading_limit`.
      oauth2MetadataUrl: http://localhost:9000/.well-known/oauth-authorization-server
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:9000/oauth2/authorize
          tokenUrl: http://localhost:9000/oauth2/token
          refreshUrl: http://localhost:9000/oauth2/token
          scopes:
            market-data:read: Instruments, quotes and option chains
            accounts:read: Accounts, balances, positions and transactions
            orders:read: Orders, previews and order events
            orders:write: Place, change and cancel orders
        clientCredentials:
          tokenUrl: http://localhost:9000/oauth2/token
          scopes:
            market-data:read: Instruments, quotes and option chains
            accounts:read: Accounts, balances, positions and transactions
            orders:read: Orders, previews and order events
            orders:write: Place, change and cancel orders
  parameters:
...
    instrumentId:
      name: instrumentId
      in: path
      required: true
      schema: { type: string, examples: [EQ-AAPL] }
...
    If-None-Match:
      name: If-None-Match
      in: header
      description: An ETag from an earlier response; an unchanged resource returns 304.
      schema: { type: string }
    API-Version:
      name: API-Version
      in: header
      description: The API version to use. Defaults to 1 (deprecated).
      schema: { type: string, enum: ["1", "2"], default: "1" }
  headers:
...
    ETag:
      description: The representation's entity tag.
      schema: { type: string }
...
    Deprecation:
      description: When this version was deprecated (RFC 9745), e.g. @1788220800.
      schema: { type: string }
    Sunset:
      description: When this version stops working (RFC 8594).
      schema: { type: string }
  responses:
    BadRequest:
      description: The request is malformed or fails validation.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          examples:
            invalidFields:
              dataValue:
                type: https://docs.brokerage.example/problems/invalid-request
                title: The request is not valid
                status: 400
                detail: One or more fields are invalid
                instance: /accounts/ACC-1001/orders
                requestId: 7f3c2a90-4a1e-4c1b-9d3e-2b8f1c6e0a11
                errors:
                  - { field: quantity, message: must be greater than 0 }
...
    NotFound:
      description: The resource does not exist, or belongs to another customer.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Decimal:
      description: An exact decimal number, as a string.
      type: string
      pattern: '^-?[0-9]+(\.[0-9]+)?$'
      examples: ["187.50"]
    Link:
      type: object
      required: [href]
      properties:
        href: { type: string }
        templated: { type: boolean }
        type: { type: string }
        name: { type: string }
        title: { type: string }
    Links:
      description: HAL links, keyed by relation. A relation may hold one link or several.
      type: object
      properties:
        self: { $ref: "#/components/schemas/Link" }
        curies:
          type: array
          items: { $ref: "#/components/schemas/Link" }
      additionalProperties:
        anyOf:
          - $ref: "#/components/schemas/Link"
          - type: array
            items: { $ref: "#/components/schemas/Link" }
...
    Problem:
      description: An RFC 9457 problem document. Extension members depend on the type.
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        requestId: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    InstrumentType:
      type: string
      enum: [EQUITY, ETF, OPTION]
    Instrument:
      type: object
      required: [id, symbol, name, type, currency, tradable, fractionable, multiplier, _links]
      properties:
        id: { type: string }
        symbol: { type: string }
        name: { type: string }
        type: { $ref: "#/components/schemas/InstrumentType" }
        exchange: { type: string, description: ISO 10383 market identifier code }
        currency: { type: string }
        tradable: { type: boolean }
        fractionable: { type: boolean }
        multiplier: { $ref: "#/components/schemas/Decimal" }
        option:
          type: object
          required: [underlyingId, right, strike, expiration]
          properties:
            underlyingId: { type: string }
            underlyingSymbol: { type: string }
            right: { type: string, enum: [CALL, PUT] }
            strike: { $ref: "#/components/schemas/Decimal" }
            expiration: { type: string, format: date }
            style: { type: string, enum: [AMERICAN, EUROPEAN] }
        _links: { $ref: "#/components/schemas/Links" }
    InstrumentCollection:
      type: object
      required: [_links]
      properties:
        _embedded:
          type: object
          properties:
            bk:instruments: { type: array, items: { $ref: "#/components/schemas/Instrument" } }
        _links: { $ref: "#/components/schemas/Links" }
    QuoteV1:
      description: Version 1 (deprecated). Flat fields.
      type: object
      required: [instrumentId, symbol, last, bid, ask, bidSize, askSize, volume, asOf]
      properties:
        instrumentId: { type: string }
        symbol: { type: string }
        last: { $ref: "#/components/schemas/Decimal" }
        bid: { $ref: "#/components/schemas/Decimal" }
        ask: { $ref: "#/components/schemas/Decimal" }
        bidSize: { $ref: "#/components/schemas/Decimal" }
        askSize: { $ref: "#/components/schemas/Decimal" }
        volume: { type: integer }
        asOf: { type: string, format: date-time }
        _links: { $ref: "#/components/schemas/Links" }
    Quote:
      description: Version 2.
      type: object
      required: [instrumentId, symbol, last, bid, ask, change, changePercent, volume, asOf]
      properties:
        instrumentId: { type: string }
        symbol: { type: string }
        last: { $ref: "#/components/schemas/Decimal" }
        bid: { $ref: "#/components/schemas/PriceLevel" }
        ask: { $ref: "#/components/schemas/PriceLevel" }
        change: { $ref: "#/components/schemas/Decimal" }
        changePercent: { $ref: "#/components/schemas/Decimal" }
        volume: { type: integer }
        asOf: { type: string, format: date-time }
        _links: { $ref: "#/components/schemas/Links" }
    PriceLevel:
      type: object
      required: [price, size]
      properties:
        price: { $ref: "#/components/schemas/Decimal" }
        size: { $ref: "#/components/schemas/Decimal" }

    # ------------------------------------------------------------------ orders
...